reformat script cleanup
