+++ scripts/events_module/dialogue/reformat/new_dialogue.json
@@ -1,0 +1,1 @@
+{}
