+++ scripts/events_module/dialogue/reformat/old_dialogue.json
@@ -1,0 +1,1 @@
+{}
