+++ scripts/events_module/dialogue/reformat_resources/all_resources.json
@@ -47,39 +47,7 @@
         "jealous": "max_respect_-20"
     },
     "file_names": [
-        "apprentice",
-        "deputy",
-        "elder",
-        "flirt",
-        "insults",
-        "former Clancat",
-        "general_no_kit",
-        "general_no_newborn",
-        "general_outsider",
-        "general_you_kit",
-        "general",
-        "kitten",
-        "kittypet",
-        "leader",
-        "loner",
-        "mediator_apprentice",
-        "mediator",
-        "medicine_cat_apprentice",
-        "medicine_cat",
-        "newborn",
-        "queen",
-        "queen's_apprentice",
-        "rogue",
-        "warrior",
-        "focuses/hailstorm",
-        "focuses/leader",
-        "focuses/starving",
-        "focuses/unknown_murder",
-        "focuses/valentines",
-        "focuses/war",
-        "focuses/halloween",
-        "focuses/april_fools",
-        "focuses/new_years"
+        "old_dialogue"
     ],
     "abbrev_list": [
         "r_c1",
